IGNITE-26478 Move dependencies
